Internally, store bound callback functions in a legit Map so they can be recalled later
Before this, the off* methods basically didn't work. The problem is that multiple calls to `.bind(...)`, on the same function, with the same input, result in different different functions. This changes implements a real Map, which uses the pre-bound function as a key, and only binds it once (which is then cached in the Map) so that there's a one-to-one relationship between callback functions and their bound counterparts.